requirements.txt
requests
flask
sqlalchemy
psycopg2-binary
pandas
python-dateutilS